On branch customRouter Extracted users validation duplicated code to a function
Changes to be committed:
modified:
  src/middlewares/validations/users.creation.mid.js
  src/middlewares/validations/users.credentilas.js
  src/utils/users.validations.js
